Move test to fuzz tests
